+++ themis.go
@@ -1,13 +1,8 @@
 // SPDX-FileCopyrightText: 2019 Comcast Cable Communications Management, LLC
 // SPDX-License-Identifier: Apache-2.0
-package main
+package themis
 
 import (
-	"errors"
-	"fmt"
-	"os"
-	"runtime"
-	"strings"
 	"time"
 
 	"github.com/InVisionApp/go-health"
@@ -23,13 +18,11 @@ import (
 	"github.com/xmidt-org/themis/xhttp/xhttpserver"
 	"github.com/xmidt-org/themis/xmetrics/xmetricshttp"
 
-	"github.com/spf13/pflag"
-	"github.com/spf13/viper"
 	"go.uber.org/fx"
 )
 
 const (
-	applicationName = "themis"
+	ApplicationName = "themis"
 )
 
 var (
@@ -38,64 +31,20 @@ var (
 	BuildTime = "undefined"
 )
 
-func Themis(opts ...fx.Option) (*fx.App, error) {
-	app := fx.New(provideAppOptions(opts...))
+func New(opts fx.Option) (*fx.App, error) {
+	app := fx.New(provideAppOptions(opts))
 
 	return app, app.Err()
 }
 
-func setupFlagSet(fs *pflag.FlagSet) error {
-	fs.StringP("file", "f", "", "the configuration file to use.  Overrides the search path.")
-	fs.Bool("dev", false, "development mode")
-	fs.String("iss", "", "the name of the issuer to put into claims.  Overrides configuration.")
-	fs.BoolP("debug", "d", false, "enables debug logging.  Overrides configuration.")
-	fs.BoolP("version", "v", false, "print version and exit")
-
-	return nil
-}
-
-func setupViper(in config.ViperIn, v *viper.Viper) (err error) {
-	if printVersion, _ := in.FlagSet.GetBool("version"); printVersion {
-		printVersionInfo()
-	}
-	if dev, _ := in.FlagSet.GetBool("dev"); dev {
-		v.SetConfigType("yaml")
-		err = v.ReadConfig(strings.NewReader(devMode))
-	} else if file, _ := in.FlagSet.GetString("file"); len(file) > 0 {
-		v.SetConfigFile(file)
-		err = v.ReadInConfig()
-	} else {
-		v.SetConfigName(string(in.Name))
-		v.AddConfigPath(fmt.Sprintf("/etc/%s", in.Name))
-		v.AddConfigPath(".")
-		v.AddConfigPath(fmt.Sprintf("$HOME/.%s", in.Name))
-		err = v.ReadInConfig()
-	}
-
-	if err != nil {
-		return
-	}
-
-	if iss, _ := in.FlagSet.GetString("iss"); len(iss) > 0 {
-		v.Set("issuer.claims.iss", iss)
-	}
-
-	if debug, _ := in.FlagSet.GetBool("debug"); debug {
-		v.Set("log.level", "DEBUG")
-	}
-
-	return
-}
-
-func provideAppOptions(opts ...fx.Option) fx.Option {
-	return fx.Options(
-		append(opts,
+func provideAppOptions(opts fx.Option) fx.Option {
+	return fx.Options(opts,
+		fx.Module(ApplicationName,
 			sallust.WithLogger(),
-			config.CommandLine{Name: applicationName}.Provide(setupFlagSet),
 			provideMetrics(),
 			token.ProvideMetrics(),
 			fx.Provide(
-				config.ProvideViper(setupViper),
+				config.ProvideViper,
 				func(u config.Unmarshaller) (c sallust.Config, err error) {
 					err = u.UnmarshalKey("log", &c)
 					return
@@ -123,14 +72,15 @@ func provideAppOptions(opts ...fx.Option) fx.Option {
 					if err != nil {
 						return candlelight.Config{}, err
 					}
-					config.ApplicationName = applicationName
+					config.ApplicationName = ApplicationName
 					return config, nil
 				},
+				fx.Private,
 			),
 			fx.Invoke(
 				xhealth.ApplyChecks(
 					&health.Config{
-						Name:     applicationName,
+						Name:     ApplicationName,
 						Interval: 24 * time.Hour,
 						Checker: xhealth.NopCheckable{
 							Details: map[string]interface{}{
@@ -147,28 +97,5 @@ func provideAppOptions(opts ...fx.Option) fx.Option {
 				BuildPprofRoutes,
 				CheckServerRequirements,
 			),
-		)...,
-	)
-}
-
-func main() {
-	app, err := Themis()
-	if errors.Is(err, pflag.ErrHelp) {
-		return
-	} else if errors.Is(err, nil) {
-		app.Run()
-	} else {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(2)
-	}
-}
-
-func printVersionInfo() {
-	fmt.Fprintf(os.Stdout, "%s:\n", applicationName)
-	fmt.Fprintf(os.Stdout, "  version: \t%s\n", Version)
-	fmt.Fprintf(os.Stdout, "  go version: \t%s\n", runtime.Version())
-	fmt.Fprintf(os.Stdout, "  built time: \t%s\n", BuildTime)
-	fmt.Fprintf(os.Stdout, "  git commit: \t%s\n", GitCommit)
-	fmt.Fprintf(os.Stdout, "  os/arch: \t%s/%s\n", runtime.GOOS, runtime.GOARCH)
-	os.Exit(0)
+		))
 }
